requirements.txt
numpy==1.26.3
pandas
scikit-surprise
fastapi
uvicorn
pyyaml
python-box
boto3
python-dotenv